--- utilities/example_data/output/components.json
@@ -1,1 +1,0 @@
-[["Correlated_2", "Correlated_1"], ["Binary_Balanced_2", "Quaternary_Balanced_2", "Quaternary_Balanced_1", "Binary_Balanced_1"], ["Continuous_Quaternary_Normal", "Continuous_Quaternary_Skewed"], ["Continuous_Trinary_Skewed", "Continuous_Trinary_Normal"], ["Trinary_Balanced_2", "Trinary_Balanced_1"], ["Quaternary_Skewed_1a", "Quaternary_Skewed_1b"]]
