+++ utilities/params.argv
@@ -1,0 +1,14 @@
+{
+  "dpi": 150,
+  "discrete_threshold": 5,
+  "debug": true,
+  "output_json": true,
+  "output_chart": true,
+  "charter": "Plotly",
+  "cache": true,
+  "sample_n": null,
+  "output_limit_n": null,
+  "input_file": "example_data/data.csv",
+  "output_dir": "example_data/output",
+  "ignore_columns":[]
+}
